code to load classifers once per machine
This introduces a new mode to run machine learning models in the
data collector. Because the ML models consume lots of RAM, we
only want to load them once per machine even if that machine
runs many DCs. The new mode can be run using 'privcount mldc',
and it has its own config file to specify the models to load
and the number of DCs to run (see test/mldc.yaml).